feat(compile-jsx): set TS
