+++ index.html
@@ -10,6 +10,6 @@
 </head>
 <body>
     <div id="root" class="test"></div>
-    <script src="./index.jsx" type="module"></script>
+    <script src="./index.tsx" type="module"></script>
 </body>
 </html>
