+++ jsx-compiler/helpers.ts
@@ -1,11 +1,12 @@
-export const createInnerText = (parent, string) => {
+import { JSX } from "./jsx";
+
+export const createInnerText = (parent: JSX.IntrinsicElements, string: string) => {
     const text = document.createTextNode(string);
     parent.appendChild(text);
 }
 
-export const createElement = (element) => {
+export const createElement = (element: JSX.IntrinsicElements) => {
     return Object.entries(element.attributes).reduce((newElement, [name, value]) => {
-        console.log(name, value)
 
         if (name.startsWith("on")) {
             newElement[name.toLowerCase()] = value;
@@ -18,7 +19,7 @@ export const createElement = (element) => {
     }, document.createElement(element.name))
 };
 
-export const isComponent = (el) => {
+export const isComponent = (el: JSX.IntrinsicElements) => {
     return el.attributes !== undefined 
         && typeof el.name === "function"
 }
