+++ jsx-compiler/jsx.ts
@@ -1,12 +1,19 @@
 const React = {
-    createElement(name, attributes = {}, ...children) {
+    createElement(name: string, attributes = {}, ...children: any[]) {
       return {
         name,
         attributes: typeof attributes === "object" && attributes !== null ? attributes : {},
         children
       };
     }
-  };
-  
+};
+
 export default React;
+
+export declare namespace JSX {
+  export interface IntrinsicElements {
+    [elemName: string]: any;
+  }
+}
+
