+++ src/components/example.tsx
@@ -1,0 +1,5 @@
+import React, {JSX} from "../../jsx-compiler/jsx";
+
+export const Example: JSX.IntrinsicElements = () => {
+  return (<div> example component </div>);
+};
